web-app/requirements.txt
aiofiles==0.5.0
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
Bottleneck==1.3.2
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
cycler==0.10.0
fastai==1.0.61
fastapi==0.56.0
fastprogress==0.2.3
future==0.18.2
h11==0.9.0
httptools==0.1.1
idna==2.9
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
multidict==4.7.6
numexpr==2.7.1
numpy==1.18.5
nvidia-ml-py3==7.352.0
packaging==20.4
pandas==1.0.4
Pillow==7.1.2
public==2019.4.13
pydantic==1.5.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-multipart==0.0.5
pytz==2020.1
PyYAML==5.3.1
query-string==2019.4.13
requests==2.23.0
scipy==1.4.1
six==1.15.0
soupsieve==2.0.1
starlette==0.13.2
urllib3==1.25.9
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
yarl==1.4.2
